Make era-test-node a library
