+++ src/formatter.rs
@@ -1,9 +1,10 @@
 //! Helper methods to display transaction data in more human readable way.
-use crate::{resolver, ShowCalls};
+use crate::resolver;
 
 use colored::Colorize;
 use serde::Deserialize;
-use std::collections::HashMap;
+use std::{collections::HashMap,  str::FromStr};
+use core::fmt::Display;
 
 use crate::fork::block_on;
 use zksync_basic_types::H160;
@@ -11,6 +12,36 @@ use zksync_basic_types::H160;
 use zksync_types::{vm_trace::Call, VmEvent};
 
 use lazy_static::lazy_static;
+
+
+#[derive(Debug, clap::Parser, Clone, clap::ValueEnum, PartialEq, Eq)]
+pub enum ShowCalls {
+    None,
+    User,
+    System,
+    All,
+}
+
+impl FromStr for ShowCalls {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_ref() {
+            "none" => Ok(ShowCalls::None),
+            "user" => Ok(ShowCalls::User),
+            "system" => Ok(ShowCalls::System),
+            "all" => Ok(ShowCalls::All),
+            _ => Err(()),
+        }
+    }
+}
+
+impl Display for ShowCalls {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "{:?}", self)
+    }
+}
+
 
 #[derive(Debug, Deserialize, Clone, PartialEq, Eq)]
 pub enum ContractType {
