+++ src/zks.rs
@@ -236,6 +236,7 @@ mod tests {
     use std::str::FromStr;
 
     use crate::node::InMemoryNode;
+    use crate::formatter::ShowCalls;
 
     use super::*;
     use zksync_basic_types::Address;
@@ -243,7 +244,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_estimate_fee() {
-        let node = InMemoryNode::new(None, crate::ShowCalls::None, false, false);
+        let node = InMemoryNode::new(None, ShowCalls::None, false, false);
         let namespace = ZkMockNamespaceImpl::new(node.get_inner());
 
         let mock_request = CallRequest {
@@ -280,7 +281,7 @@ mod tests {
     #[tokio::test]
     async fn test_get_token_price_given_eth_should_return_price() {
         // Arrange
-        let node = InMemoryNode::new(None, crate::ShowCalls::None, false, false);
+        let node = InMemoryNode::new(None, ShowCalls::None, false, false);
         let namespace = ZkMockNamespaceImpl::new(node.get_inner());
 
         let mock_address = Address::from_str("0x0000000000000000000000000000000000000000")
@@ -296,7 +297,7 @@ mod tests {
     #[tokio::test]
     async fn test_get_token_price_given_capitalized_link_address_should_return_price() {
         // Arrange
-        let node = InMemoryNode::new(None, crate::ShowCalls::None, false, false);
+        let node = InMemoryNode::new(None, ShowCalls::None, false, false);
         let namespace = ZkMockNamespaceImpl::new(node.get_inner());
 
         let mock_address = Address::from_str("0x40609141Db628BeEE3BfAB8034Fc2D8278D0Cc78")
@@ -312,7 +313,7 @@ mod tests {
     #[tokio::test]
     async fn test_get_token_price_given_unknown_address_should_return_error() {
         // Arrange
-        let node = InMemoryNode::new(None, crate::ShowCalls::None, false, false);
+        let node = InMemoryNode::new(None, ShowCalls::None, false, false);
         let namespace = ZkMockNamespaceImpl::new(node.get_inner());
 
         let mock_address = Address::from_str("0x0000000000000000000000000000000000000042")
